pyproject.toml
[tool.poetry]
name = "drift-study"
version = "0.1.0"
description = ""
authors = ["Thibault Simonetto <thibault.simonetto.001@student.uni.lu>"]

[tool.poetry.dependencies]
python = "3.8.10"
scikit-learn = "1.0.2"
tqdm = "^4.64.0"
h5py = "^3.6.0"
river = "^0.10.1"
evidently = "^0.1.50-alpha.0"
serval-config-utils = "^0.1.4"
serval-ml-commons = "0.1.0"
tables = "^3.7.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^3.0.0"
pre-commit = "^2.19.0"
flake8 = "^4.0.1"
mypy = "^0.950"
isort = "^5.10.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
